Speech2Text Analyser/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.2
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asgiref==3.7.2
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
backoff==2.2.1
bcrypt==4.1.2
bertopic==0.16.3
blinker==1.8.2
build==1.0.3
cachetools==5.3.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
ctranslate2==4.3.1
cycler==0.12.1
Cython==0.29.37
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
einops==0.8.0
exceptiongroup==1.2.0
fastapi==0.109.2
faster-whisper==1.0.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.28.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
groq==0.9.0
grpcio==1.65.1
h11==0.14.0
h5py==3.11.0
hdbscan==0.8.37
httpcore==1.0.3
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.7
importlib-metadata==6.11.0
importlib-resources==6.1.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
julius==0.2.7
keras==3.4.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.7
langchain-community==0.0.20
langchain-core
langchain-groq
langsmith==0.1.93
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lightning==2.3.3
lightning-utilities==0.11.5
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
mmh3==4.1.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.17.0
openai==1.12.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
optree==0.12.1
optuna==3.6.1
orjson==3.10.6
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
pooch==1.8.2
posthog==3.4.1
primePy==1.3
protobuf==4.25.3
pulsar-client==3.4.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytorch-lightning==2.3.3
pytorch-metric-learning==2.5.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
seaborn==0.13.2
semver==3.0.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
SQLAlchemy==2.0.27
starlette==0.36.3
streamlit==1.37.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-hub==0.16.1
termcolor==2.4.0
tf_keras==2.17.0
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
top2vec==1.0.34
torch==2.3.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.3.1
torchmetrics==1.4.0.post0
tornado==6.4.1
tqdm==4.66.2
transformers==4.39.3
urllib3==2.2.2
whisperx==3.1.5
wordcloud==1.9.3
